+++ util/util_misc.h
@@ -23,6 +23,7 @@ SOFTWARE.
 #ifndef __UTIL_MISC_H__
 #define __UTIL_MISC_H__
 
+#include <stdbool.h>
 #include <sys/types.h>
 #include <sys/socket.h>
 #include <netinet/in.h>
@@ -128,37 +129,37 @@ float hypotenuse(float x, float y, float z);
 
 // -----------------  SMOOTHING  -----------------------------------------
 
-void basic_exponential_smoothing(long double x, long double *s, long double alpha);
-void double_exponential_smoothing(long double x, long double *s, long double *b, long double alpha, long double beta, bool init);
+void basic_exponential_smoothing(double x, double *s, double alpha);
+void double_exponential_smoothing(double x, double *s, double *b, double alpha, double beta, bool init);
 
 // -----------------  MOVING AVERAGE  ------------------------------------
 
 typedef struct {
     int64_t max_values;
-    long double sum;
+    double sum;
     int64_t count;
-    long double current;
-    long double values[0];
+    double current;
+    double values[0];
 } ma_t;
-long double moving_average(long double val, ma_t *ma);
-long double moving_average_query(ma_t *ma);
+double moving_average(double val, ma_t *ma);
+double moving_average_query(ma_t *ma);
 ma_t * moving_average_alloc(int32_t max_values);
 void moving_average_free(ma_t * ma) ;
 void moving_average_reset(ma_t * ma);
 
 typedef struct {
-    long double time_span;
+    double time_span;
     int64_t max_bins;
     ma_t * ma;
     bool first_call;
     int64_t last_idx;
-    long double sum;
+    double sum;
     int64_t count;
-    long double current;
+    double current;
 } tma_t;
-long double timed_moving_average(long double val, long double time_arg, tma_t *tma);
-long double timed_moving_average_query(tma_t *tma);
-tma_t * timed_moving_average_alloc(long double time_span, int64_t max_bins);
+double timed_moving_average(double val, double time_arg, tma_t *tma);
+double timed_moving_average_query(tma_t *tma);
+tma_t * timed_moving_average_alloc(double time_span, int64_t max_bins);
 void timed_moving_average_free(tma_t * tma);
 void timed_moving_average_reset(tma_t * tma);
